+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. The format 
 
 - Add `queryComponent`, `queryComponentAll` and `closestComponent` helper functions ([#629](https://github.com/studiometa/js-toolkit/issues/629), [#669](https://github.com/studiometa/js-toolkit/pull/669))
 - Add `$query(name)` and `$closest(name)` methods on `Base` ([#694](https://github.com/studiometa/js-toolkit/issues/694))
+- Add `defineFeatures()` helper to configure features without `createApp` ([#691](https://github.com/studiometa/js-toolkit/issues/691))
 
 ## [v3.4.3](https://github.com/studiometa/js-toolkit/compare/3.4.2..3.4.3) (2025-11-13)